# opaa-api/src/main/resources/openapi/opaa-api.yaml
openapi: 3.0.3
info:
  title: OPAA API
  description: Open Project AI Assistant REST API
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
paths:
...
  /api/v1/libraries:
    get:
      operationId: listLibraries
      tags: [library]
      summary: List knowledge libraries the current user holds rights on - direct grant, grant to one of their groups, or organization-wide visibility (ownership is included, since creating a library grants the creator OWNER)
      responses:
        "200":
          description: List of libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createLibrary
      tags: [library]
      summary: Create a new knowledge library, owned by the caller or by one of the caller's groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryRequest"
      responses:
        "201":
          description: Library created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/libraries/source-test:
    post:
      operationId: testLibrarySource
      tags: [library]
      summary: >
        Test a quellkonfiguration before a library is created, or an existing library's
        quellkonfiguration while editing it (#514, #544). Without libraryId, the permission bar
        is the same as createLibrary (any authenticated user; there is no library yet to hold a
        role on). With libraryId, the caller additionally needs AssetRole.MANAGER on that
        library (see SourceConnectionTestRequest.libraryId).
      description: >
        Checks the given sourceType/sourcePath/sourceUrl/sourceProxy/sourceCredentials/
        sourceInsecureSsl combination the way the corresponding indexing run later would,
        using the same HTTP client building blocks (io.opaa.sourceaccess.SourceHttpClientFactory.buildHttpClient)
        and, for FILESYSTEM, the same operator-configured allowlist (FilesystemPathAllowlist)
        createLibrary itself enforces - a path outside the allowlist is rejected with 400
        before anything on disk is touched, exactly like creation. sourceType UPLOAD has
        nothing to test and is rejected with 400. The response never reveals more about a
        FILESYSTEM directory than a count, and every failure message is German and intended
        to be shown to the caller as-is. HTTP_DIRECTORY/RSS_FEED response bodies are bounded
        the same way the RSS indexing run bounds them (opaa.indexing.rss.max-page-size-bytes/
        max-feed-size-bytes) - an unbounded read here would let a caller crash the server with
        a single oversized or endless response. This endpoint is rate-limited per IP and
        globally (opaa.rate-limit.source-test), same as the indexing trigger - it is a
        synchronous probe with its own outbound connections and short timeouts, not a
        replacement for #267's still-outstanding target validation (blocking internal/private
        address ranges) for HTTP_DIRECTORY/RSS_FEED, which this endpoint makes cheaper to
        reach than the indexing run it mirrors (see docs/features/knowledge-sources.md). With
        libraryId set (#544), sourceType must match that library's own sourceType (otherwise
        400) and a missing sourceCredentials falls back to the library's stored one, but only
        if sourceUrl still names the same origin as the library's stored sourceUrl - the same
        rule KnowledgeLibraryService#validateSourceConfigurationForUpdate already applies when
        saving.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceConnectionTestRequest"
      responses:
        "200":
          description: >
            The test ran to completion - reachable distinguishes success from a diagnosed
            failure (see message); this is not itself an error response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceConnectionTestResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/libraries/confluence/spaces:
    post:
      operationId: listConfluenceSpaces
      tags: [library]
      summary: >
        List the spaces a Confluence token may read, for the space selection of a CONFLUENCE
        library (ADR-0023, #1134).
      description: >
        Authenticates against the instance with the given credentials (Cloud: e-mail + API
        token as HTTP Basic against /wiki/api/v2; Data Center: personal access token as Bearer
        against /rest/api) and returns every readable space, following the instance's own
        pagination to the end. A POST because the credentials travel in the body and no library
        needs to exist yet. Permission bar like the connection test: any authenticated user
        without libraryId, AssetRole.MANAGER on that library with it. Rate-limited like the
        connection test (opaa.rate-limit.source-test) and subject to the same outbound target
        validation - a private on-premises address needs its allowlist entry and the error
        message says so. Credentials appear in no response and no error message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfluenceSpaceListRequest"
      responses:
        "200":
          description: The readable spaces.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfluenceSpaceListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/libraries/{libraryId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: getLibrary
      tags: [library]
      summary: Get library details
      responses:
        "200":
          description: Library details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateLibrary
      tags: [library]
      summary: Update library name, description, visibility and listed flag (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryUpdateRequest"
      responses:
        "200":
          description: Library updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteLibrary
      tags: [library]
      summary: >
        Delete a library (owner only). A non-empty UPLOAD
        library is rejected with 409 (documents must be removed individually first). A non-empty
        connector library (FILESYSTEM/HTTP_DIRECTORY/RSS_FEED) is deleted together with its
        documents and vector store chunks - deleting individual connector documents would be
        undone by the next indexing run (ADR-0018, Entscheidung 5). The delete is also rejected
        with 409 while an indexing run for the library is RUNNING.
      responses:
        "204":
          description: Library deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/libraries/{libraryId}/documents/{documentId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      operationId: deleteLibraryDocument
      tags: [library]
      summary: >
        Remove a document from a library (EDITOR role or above required): the document row, its
        chunks in the vector store, and the stored file.
      responses:
        "204":
          description: Document deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/libraries/{libraryId}/folders:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      operationId: createLibraryFolder
      tags: [library]
      summary: >
        Create a folder in a library (EDITOR role or above required), optionally nested under an
        existing folder (#820, ADR-0020: library_folders is a real entity so an empty folder can
        exist before any document is uploaded into it). Only accepted for libraries with sourceType
        UPLOAD - a connector library (FILESYSTEM/HTTP_DIRECTORY/RSS_FEED) rejects folder creation
        with 409, mirroring the same restriction POST .../documents already applies (ADR-0018,
        #479); FILESYSTEM's own read-only folder mapping is a later, separate task (#824). A name
        that already exists on the same level (same parentFolderId, including two root-level
        folders) is rejected with 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LibraryFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryFolderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/libraries/{libraryId}/grants:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      operationId: listAssetGrants
      tags: [library]
      summary: >
        List the grants on a library (MANAGER role or above required). See
        docs/features/spaces-and-assets.md#asset-rollen.
      responses:
        "200":
          description: List of grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssetGrantResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: upsertAssetGrant
      tags: [library]
      summary: >
        Grant (or update the role of) a user or group on a library (MANAGER role or above
        required). Idempotent per subject: submitting again for the same subject replaces its role
        and expiry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetGrantRequest"
      responses:
        "200":
          description: Grant created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetGrantResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/libraries/{libraryId}/grants/{grantId}:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: grantId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      operationId: revokeAssetGrant
      tags: [library]
      summary: Revoke a grant (MANAGER role or above required). Takes effect on the next query.
      responses:
        "204":
          description: Grant revoked
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  schemas:
...
    ConfluenceSpaceListRequest:
      type: object
      required: [sourceUrl, confluenceEdition]
      description: >
        Connection data for listing the spaces a Confluence token may read (ADR-0023; the basis
        of the wizard's space selection). Credentials are used for this one request only, never
        persisted or echoed. With libraryId (an existing CONFLUENCE library the caller manages),
        omitted credentials fall back to the stored ones when sourceUrl names the same origin -
        the same rule the connection test applies.
      properties:
        sourceUrl:
          type: string
          format: uri
          description: The instance's base address, normalised like LibraryRequest.sourceUrl.
        confluenceEdition:
          $ref: "#/components/schemas/ConfluenceEdition"
        sourceCredentials:
          type: string
          maxLength: 500
          nullable: true
          description: Credentials in the edition's format (see LibraryRequest.sourceCredentials).
        sourceProxy:
          type: string
          maxLength: 255
          nullable: true
        sourceInsecureSsl:
          type: boolean
          nullable: true
          default: false
        libraryId:
          type: string
          format: uuid
          nullable: true
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...